Update respone API

+++ src/main/java/net/javaguides/class_student/serviceImp/StudentServiceImp.java
@@ -3,40 +3,60 @@ package net.javaguides.class_student.serviceImp;
 import net.javaguides.class_student.model.Student;
 import net.javaguides.class_student.repository.StudentRepository;
 import net.javaguides.class_student.service.StudentService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
 public class StudentServiceImp implements StudentService {
 
+    @Autowired
     private StudentRepository studentRepository;
+
     @Override
-    public List<Student> getAllStudent() {
-        return studentRepository.findAll();
+    public ResponseEntity<List<Student>> getAllStudent() {
+        List<Student> listStudent = studentRepository.findAll();
+        if(listStudent.isEmpty()) {
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity(listStudent, HttpStatus.OK);
     }
 
     @Override
-    public String createStudent(Student student) {
-        studentRepository.save(student);
-        return "Create success";
+    public Student createStudent(Student input) {
+        return studentRepository.save(input);
     }
 
     @Override
-    public String updateStudentById(long id, Student student) {
-        Student data = studentRepository.findById(id).get();
+    public ResponseEntity<Student> updateStudentById(long id, Student input) {
+        Student student = studentRepository.findById(id).get();
 
-        data.setFullName(data.getFullName());
-        data.setAge(data.getAge());
-        data.setSex(data.getSex());
-        data.setCountry(data.getCountry());
+        if(student == null) {
+            return ResponseEntity.notFound().build();
+        }
 
-        studentRepository.save(data);
+        student.setFullName(input.getFullName());
+        student.setAge(input.getAge());
+        student.setSex(input.getSex());
+        student.setCountry(input.getCountry());
 
-        return "updated success";
+        Student updatedStudent = studentRepository.save(student);
+
+        return ResponseEntity.ok(updatedStudent);
     }
 
     @Override
-    public String deleteStudentById(long id) {
+    public ResponseEntity<Student> deleteStudentById(long id) {
+        Student student = studentRepository.findById(id).get();
+
+        if(student == null) {
+            return ResponseEntity.notFound().build();
+        }
+
         studentRepository.deleteById(id);
-        return "delete success";
+        return ResponseEntity.ok().build();
     }
 }
